Refactor modal actions into their own files and simplify logic
