+++ src/editor/actions/actions.js
@@ -1,15 +1,5 @@
 import { getUrlParams, objectToQueryString } from "../tools/query-param-tools";
-import { getRevisionIdsNeededForDisplay } from "../tools/revision-history";
-import {
-  getRevisionList,
-  getRevisions
-} from "../../shared/server-api/revisions";
-import { haveLocalAutosave } from "../tools/local-autosave";
-import {
-  getNotebookID,
-  getUserDataFromDocument,
-  isLoggedIn
-} from "../tools/server-tools";
+import { getNotebookID, getUserDataFromDocument } from "../tools/server-tools";
 
 import { getChunkContainingLine } from "../iomd-tools/iomd-selection";
 
@@ -116,112 +106,10 @@ export function moveCursorToNextChunk() {
   };
 }
 
-function getRequiredRevisionContent(state, dispatch) {
-  const contentIdsNeeded = getRevisionIdsNeededForDisplay(
-    state.notebookHistory
-  );
-
-  // if we don't need anything, just return here!!
-  if (!contentIdsNeeded.length) {
-    return;
-  }
-
-  dispatch({
-    type: "GETTING_REVISION_CONTENT"
-  });
-  getRevisions(getNotebookID(state), contentIdsNeeded, isLoggedIn(state))
-    .then(revisions => {
-      // reduce the revisions array into an object whose keys
-      // are revision ids, and whose body is the content of
-      // the revisions
-      const revisionContent = revisions.reduce(
-        (acc, r) => Object.assign(acc, { [r.id]: r.content }),
-        {}
-      );
-      dispatch({
-        type: "GOT_REVISION_CONTENT",
-        revisionContent
-      });
-    })
-    .catch(() => {
-      dispatch({ type: "ERROR_GETTING_REVISION_CONTENT" });
-    });
-}
-
-export function updateSelectedRevisionId(selectedRevisionId) {
-  return (dispatch, getState) => {
-    dispatch({
-      type: "UPDATE_NOTEBOOK_HISTORY_BROWSER_SELECTED_REVISION_ID",
-      selectedRevisionId
-    });
-    getRequiredRevisionContent(getState(), dispatch);
-  };
-}
-
-export function getNotebookRevisionList() {
-  return (dispatch, getState) => {
-    dispatch({ type: "GETTING_NOTEBOOK_REVISION_LIST" });
-    getRevisionList(getNotebookID(getState()), isLoggedIn(getState()))
-      .then(revisionList => {
-        haveLocalAutosave(getState())
-          .then(havePendingChanges => {
-            dispatch({
-              type: "UPDATE_NOTEBOOK_HISTORY",
-              hasLocalOnlyChanges: havePendingChanges,
-              revisionList,
-              selectedRevisionId: havePendingChanges
-                ? undefined
-                : revisionList[0].id
-            });
-            getRequiredRevisionContent(getState(), dispatch);
-          })
-          .catch(() => {
-            dispatch({ type: "ERROR_GETTING_NOTEBOOK_REVISION_LIST" });
-          });
-      })
-      .catch(() => {
-        dispatch({ type: "ERROR_GETTING_NOTEBOOK_REVISION_LIST" });
-      });
-  };
-}
-
-export function setModalState(modalState) {
-  return {
-    type: "SET_MODAL_STATE",
-    modalState
-  };
-}
-
 export function updateNotebookInfo(notebookInfo) {
   return {
     type: "UPDATE_NOTEBOOK_INFO",
     notebookInfo
-  };
-}
-
-export function toggleHistoryModal() {
-  return (dispatch, getState) => {
-    const modalState =
-      getState().modalState === "HISTORY_MODAL"
-        ? "MODALS_CLOSED"
-        : "HISTORY_MODAL";
-    dispatch(setModalState(modalState));
-  };
-}
-
-export function toggleHelpModal() {
-  return (dispatch, getState) => {
-    const modalState =
-      getState().modalState === "HELP_MODAL" ? "MODALS_CLOSED" : "HELP_MODAL";
-    dispatch(setModalState(modalState));
-  };
-}
-
-export function toggleFileModal() {
-  return (dispatch, getState) => {
-    const modalState =
-      getState().modalState === "FILE_MODAL" ? "MODALS_CLOSED" : "FILE_MODAL";
-    dispatch(setModalState(modalState));
   };
 }
